initial query pools from node
